chore: move FeeRate component from product-components to suite
