+++ packages/product-components/src/index.ts
@@ -20,7 +20,6 @@ export { NumberInput } from './components/NumberInput/NumberInput';
 export { InputWithOptions } from './components/InputWithOptions/InputWithOptions';
 export { EditableText, type EditableTextProps } from './components/EditableText/EditableText';
 export { CardButton } from './components/CardButton/CardButton';
-export { FeeRate } from './components/FeeRate/FeeRate';
 export {
     DeviceAnimation,
     type DeviceAnimationProps,
