+++ packages/suite/src/components/wallet/FeeRate/FeeRate.tsx
